+++ app/api/__init__.py
@@ -1,0 +1,5 @@
+from .rest import Restful
+from .pgs import Pgs
+
+__all__ = ['Restful', 'Pgs'] 
+